Added additional file types
Updated script to work with multiple types of text files.

'send_log' is no longer required to send data to Humio, but is used as a 'catch all' when Humio transmission fails and results are left in the RTR directory, or when you run a script that
only provides a local log file for output (like 'run_cli_tool').