Update send_log.ps1
Updated to correct bugs related to the secondary PowerShell processes that are launched to wait for the file to be accessible.

If a file is currently locked, a second PowerShell process will start and check every 30 seconds, for a total of 5 minutes, for the file to be unlocked. If the file is unlocked during that period, it will be picked up and sent to Humio.